Move each view function to own file
