+++ codelists/views/__init__.py
@@ -1,0 +1,9 @@
+from .codelist import codelist
+from .codelist_create import CodelistCreate
+from .codelist_update import CodelistUpdate
+from .index import index
+from .version import version
+from .version_create import VersionCreate
+from .version_download import version_download
+from .version_publish import version_publish
+from .version_update import VersionUpdate
