+++ codelists/views/version_download.py
@@ -1,0 +1,31 @@
+from django.http import Http404, HttpResponse
+from django.shortcuts import get_object_or_404
+
+from ..models import CodelistVersion
+
+
+def version_download(request, organisation_slug, codelist_slug, qualified_version_str):
+    if qualified_version_str[-6:] == "-draft":
+        expect_draft = True
+        version_str = qualified_version_str[:-6]
+    else:
+        expect_draft = False
+        version_str = qualified_version_str
+
+    clv = get_object_or_404(
+        CodelistVersion.objects.select_related("codelist"),
+        codelist__organisation_id=organisation_slug,
+        codelist__slug=codelist_slug,
+        version_str=version_str,
+    )
+
+    if expect_draft != clv.is_draft:
+        raise Http404
+
+    response = HttpResponse(content_type="text/csv")
+    content_disposition = 'attachment; filename="{}.csv"'.format(
+        clv.download_filename()
+    )
+    response["Content-Disposition"] = content_disposition
+    response.write(clv.csv_data)
+    return response
